Move map format loading and saving into mods.
