+++ OpenRA.Game/ModData.cs
@@ -31,6 +31,7 @@ namespace OpenRA
 		public readonly ISpriteLoader[] SpriteLoaders;
 		public readonly ISpriteSequenceLoader SpriteSequenceLoader;
 		public readonly IModelSequenceLoader ModelSequenceLoader;
+		public readonly IMapLoader MapLoader;
 		public readonly HotkeyManager Hotkeys;
 		public ILoadScreen LoadScreen { get; private set; }
 		public CursorProvider CursorProvider { get; private set; }
@@ -89,6 +90,8 @@ namespace OpenRA
 
 			ModelSequenceLoader = (IModelSequenceLoader)modelCtor.Invoke(new[] { this });
 			ModelSequenceLoader.OnMissingModelError = s => Log.Write("debug", s);
+
+			MapLoader = Manifest.Get<IMapLoader>();
 
 			Hotkeys = new HotkeyManager(ModFiles, Game.Settings.Keys, Manifest);
 
@@ -185,7 +188,7 @@ namespace OpenRA
 
 			Map map;
 			using (new Support.PerfTimer("Map"))
-				map = new Map(this, MapCache[uid].Package);
+				map = MapLoader.Load(this, MapCache[uid].Package);
 
 			LoadTranslations(map);
