+++ OpenRA.Mods.Common/UtilityCommands/ResizeMapCommand.cs
@@ -49,7 +49,7 @@ namespace OpenRA.Mods.Common.UtilityCommands
 		void IUtilityCommand.Run(Utility utility, string[] args)
 		{
 			var modData = Game.ModData = utility.ModData;
-			map = new Map(modData, new Folder(".").OpenPackage(args[1], modData.ModFiles));
+			map = modData.MapLoader.Load(modData, new Folder(".").OpenPackage(args[1], modData.ModFiles));
 			Console.WriteLine("Resizing map {0} from {1} to {2},{3}", map.Title, map.MapSize, width, height);
 			map.Resize(width, height);
