+++ OpenRA.Mods.Common/UtilityCommands/Utilities.cs
@@ -32,7 +32,7 @@ namespace OpenRA.Mods.Common.UtilityCommands
 			{
 				try
 				{
-					map = new Map(modData, new Folder(".").OpenPackage(mapPath, modData.ModFiles));
+					map = modData.MapLoader.Load(modData, new Folder(".").OpenPackage(mapPath, modData.ModFiles));
 				}
 				catch (InvalidDataException ex)
 				{
